requirements.txt
Django>=5.1.6
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.28.0
